+++ shenyu-plugin/pom.xml
@@ -55,6 +55,6 @@
         <module>shenyu-plugin-logging</module>
         <module>shenyu-plugin-request</module>
         <module>shenyu-plugin-jwt</module>
-        <module>shenyu-plugin-rpc-response</module>
+        <module>shenyu-plugin-response</module>
     </modules>
 </project>
