+++ shenyu-plugin/shenyu-plugin-alibaba-dubbo/src/main/java/org/apache/shenyu/plugin/alibaba/dubbo/AlibabaDubboPlugin.java
@@ -76,9 +76,9 @@ public class AlibabaDubboPlugin extends AbstractShenyuPlugin {
         }
         Object result = alibabaDubboProxyService.genericInvoker(param, metaData);
         if (Objects.nonNull(result)) {
-            exchange.getAttributes().put(Constants.DUBBO_RPC_RESULT, result);
+            exchange.getAttributes().put(Constants.RPC_RESULT, result);
         } else {
-            exchange.getAttributes().put(Constants.DUBBO_RPC_RESULT, Constants.DUBBO_RPC_RESULT_EMPTY);
+            exchange.getAttributes().put(Constants.RPC_RESULT, Constants.DUBBO_RPC_RESULT_EMPTY);
         }
         exchange.getAttributes().put(Constants.CLIENT_RESPONSE_RESULT_TYPE, ResultEnum.SUCCESS.getName());
         return chain.execute(exchange);
