+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-plugin/pom.xml
@@ -53,6 +53,6 @@
         <module>shenyu-spring-boot-starter-plugin-logging</module>
         <module>shenyu-spring-boot-starter-plugin-request</module>
         <module>shenyu-spring-boot-starter-plugin-jwt</module>
-        <module>shenyu-spring-boot-starter-plugin-rpc-response</module>
+        <module>shenyu-spring-boot-starter-plugin-response</module>
     </modules>
 </project>
